--- include/document/SemanticTokenTypes.hpp
@@ -1,152 +1,0 @@
-#pragma once
-
-#include <cstdint>
-#include <string_view>
-#include <vector>
-
-namespace lsp {
-
-enum class SemanticTokenType : uint32_t {
-    Namespace = 0,
-    Type = 1,
-    Class = 2,
-    Enum = 3,
-    Interface = 4,
-    Struct = 5,
-    TypeParameter = 6,
-    Parameter = 7,
-    Variable = 8,
-    Property = 9,
-    EnumMember = 10,
-    Event = 11,
-    Function = 12,
-    Method = 13,
-    Macro = 14,
-    Keyword = 15,
-    Modifier = 16,
-    Comment = 17,
-    String = 18,
-    Number = 19,
-    Regexp = 20,
-    Operator = 21,
-    Decorator = 22,
-};
-
-enum class SemanticTokenModifier : uint32_t {
-    Declaration = 0,
-    Definition = 1,
-    Readonly = 2,
-    Static = 3,
-    Deprecated = 4,
-    Abstract = 5,
-    Async = 6,
-    Modification = 7,
-    Documentation = 8,
-    DefaultLibrary = 9,
-};
-
-inline constexpr std::string_view toString(SemanticTokenType type) {
-    switch (type) {
-        case SemanticTokenType::Namespace:
-            return "namespace";
-        case SemanticTokenType::Type:
-            return "type";
-        case SemanticTokenType::Class:
-            return "class";
-        case SemanticTokenType::Enum:
-            return "enum";
-        case SemanticTokenType::Interface:
-            return "interface";
-        case SemanticTokenType::Struct:
-            return "struct";
-        case SemanticTokenType::TypeParameter:
-            return "typeParameter";
-        case SemanticTokenType::Parameter:
-            return "parameter";
-        case SemanticTokenType::Variable:
-            return "variable";
-        case SemanticTokenType::Property:
-            return "property";
-        case SemanticTokenType::EnumMember:
-            return "enumMember";
-        case SemanticTokenType::Event:
-            return "event";
-        case SemanticTokenType::Function:
-            return "function";
-        case SemanticTokenType::Method:
-            return "method";
-        case SemanticTokenType::Macro:
-            return "macro";
-        case SemanticTokenType::Keyword:
-            return "keyword";
-        case SemanticTokenType::Modifier:
-            return "modifier";
-        case SemanticTokenType::Comment:
-            return "comment";
-        case SemanticTokenType::String:
-            return "string";
-        case SemanticTokenType::Number:
-            return "number";
-        case SemanticTokenType::Regexp:
-            return "regexp";
-        case SemanticTokenType::Operator:
-            return "operator";
-        case SemanticTokenType::Decorator:
-            return "decorator";
-    }
-
-    return "";
-}
-
-inline constexpr std::string_view toString(SemanticTokenModifier modifier) {
-    switch (modifier) {
-        case SemanticTokenModifier::Declaration:
-            return "declaration";
-        case SemanticTokenModifier::Definition:
-            return "definition";
-        case SemanticTokenModifier::Readonly:
-            return "readonly";
-        case SemanticTokenModifier::Static:
-            return "static";
-        case SemanticTokenModifier::Deprecated:
-            return "deprecated";
-        case SemanticTokenModifier::Abstract:
-            return "abstract";
-        case SemanticTokenModifier::Async:
-            return "async";
-        case SemanticTokenModifier::Modification:
-            return "modification";
-        case SemanticTokenModifier::Documentation:
-            return "documentation";
-        case SemanticTokenModifier::DefaultLibrary:
-            return "defaultLibrary";
-    }
-
-    return "";
-}
-
-inline std::vector<std::string_view> semanticTokenTypeLegend() {
-    return {
-        "namespace",     "type",      "class",    "enum",     "interface",  "struct",
-        "typeParameter", "parameter", "variable", "property", "enumMember", "event",
-        "function",      "method",    "macro",    "keyword",  "modifier",   "comment",
-        "string",        "number",    "regexp",   "operator", "decorator",
-    };
-}
-
-inline std::vector<std::string_view> semanticTokenModifierLegend() {
-    return {
-        "declaration", "definition", "readonly",     "static",        "deprecated",
-        "abstract",    "async",      "modification", "documentation", "defaultLibrary",
-    };
-}
-
-constexpr uint32_t semanticTokenTypeIndex(SemanticTokenType type) {
-    return static_cast<uint32_t>(type);
-}
-
-constexpr uint32_t semanticTokenModifierMask(SemanticTokenModifier modifier) {
-    return 1u << static_cast<uint32_t>(modifier);
-}
-
-} // namespace lsp
